2
this is the first time that i'd like to write a review of the code of leetcode's problem. i want to insist on it.
Spending a lot of submission times on it, i have solved it finally.
though the solution is general rough, it can still bring experience to me.